+++ compiler/src/org.graalvm.compiler.truffle.runtime/src/org/graalvm/compiler/truffle/runtime/OptimizedCallTarget.java
@@ -70,7 +70,7 @@ import jdk.vm.ci.meta.SpeculationLog;
  * Note: {@code PartialEvaluator} looks up this class and a number of its methods by name.
  */
 @SuppressWarnings("deprecation")
-public abstract class OptimizedCallTarget implements CompilableTruffleAST, RootCallTarget, ReplaceObserver, com.oracle.truffle.api.LoopCountReceiver {
+public abstract class OptimizedCallTarget implements CompilableTruffleAST, RootCallTarget, ReplaceObserver {
 
     private static final String NODE_REWRITING_ASSUMPTION_NAME = "nodeRewritingAssumption";
     static final String CALL_BOUNDARY_METHOD_NAME = "callProxy";
@@ -131,6 +131,7 @@ public abstract class OptimizedCallTarget implements CompilableTruffleAST, RootC
         uninitializedNodeCount = runtime().getTvmci().adoptChildrenAndCount(this.rootNode);
         RuntimeOptionsCache.reinitialize();
         knownCallNodes = RuntimeOptionsCache.isLegacySplitting() ? null : new ArrayList<>(1);
+        runtime().getTvmci().setCallTarget(rootNode, this);
     }
 
     public Assumption getNodeRewritingAssumption() {
@@ -563,14 +564,6 @@ public abstract class OptimizedCallTarget implements CompilableTruffleAST, RootC
     }
 
     final void onLoopCount(int count) {
-        getCompilationProfile().reportLoopCount(count);
-    }
-
-    /*
-     * For compatibility of Graal runtime with older Truffle runtime. Remove after 0.12.
-     */
-    @Override
-    public void reportLoopCount(int count) {
         getCompilationProfile().reportLoopCount(count);
     }