+++ compiler/src/org.graalvm.compiler.truffle.test/src/org/graalvm/compiler/truffle/test/TruffleInliningTest.java
@@ -220,8 +220,7 @@ public abstract class TruffleInliningTest {
                     OptimizedDirectCallNode callNode = new OptimizedDirectCallNode(target);
                     callSites.add(callNode);
                     for (int i = 0; i < instruction.count; i++) {
-                        Integer[] args = {0};
-                        callNode.call(args);
+                        callNode.call(0);
                     }
                 }
                 InlineTestRootNode rootNode = (InlineTestRootNode) caller.getRootNode();