+++ truffle/src/com.oracle.truffle.api.utilities/src/com/oracle/truffle/api/utilities/JSONHelper.java
@@ -55,13 +55,9 @@ import com.oracle.truffle.api.nodes.RootNode;
  *
  * @since 0.8 or earlier
  */
-public class JSONHelper {
-    /**
-     * @deprecated accidentally public - to be removed.
-     * @since 0.8 or earlier
-     */
-    @Deprecated
-    public JSONHelper() {
+public final class JSONHelper {
+
+    private JSONHelper() {
     }
 
     private static StringBuilder AstJsonDumpBuilder = new StringBuilder();
@@ -175,12 +171,8 @@ public class JSONHelper {
      * @since 0.8 or earlier
      */
     public abstract static class JSONStringBuilder {
-        /**
-         * @deprecated accidentally public - don't use
-         * @since 0.8 or earlier
-         */
-        @Deprecated
-        protected JSONStringBuilder() {
+
+        private JSONStringBuilder() {
         }
 
         /** @since 0.8 or earlier */