+++ truffle/src/com.oracle.truffle.api/src/com/oracle/truffle/api/impl/Accessor.java
@@ -103,6 +103,8 @@ public abstract class Accessor {
 
         public abstract boolean isInstrumentable(RootNode rootNode);
 
+        public abstract void setCallTarget(RootNode rootNode, RootCallTarget callTarget);
+
         public abstract boolean isTaggedWith(Node node, Class<?> tag);
 
         public abstract boolean isCloneUninitializedSupported(RootNode rootNode);
@@ -729,16 +731,6 @@ public abstract class Accessor {
     protected void onLoopCount(Node source, int iterations) {
         if (SUPPORT != null) {
             SUPPORT.onLoopCount(source, iterations);
-        } else {
-            // needs an additional compatibility check so older graal runtimes
-            // still run with newer truffle versions
-            RootNode root = source.getRootNode();
-            if (root != null) {
-                RootCallTarget target = root.getCallTarget();
-                if (target instanceof com.oracle.truffle.api.LoopCountReceiver) {
-                    ((com.oracle.truffle.api.LoopCountReceiver) target).reportLoopCount(iterations);
-                }
-            }
         }
     }