+++ truffle/src/com.oracle.truffle.api/src/com/oracle/truffle/api/nodes/IndirectCallNode.java
@@ -41,9 +41,7 @@
 package com.oracle.truffle.api.nodes;
 
 import com.oracle.truffle.api.CallTarget;
-import com.oracle.truffle.api.CompilerDirectives.CompilationFinal;
 import com.oracle.truffle.api.Truffle;
-import com.oracle.truffle.api.frame.VirtualFrame;
 
 /**
  * Represents an indirect call to a {@link CallTarget}. Indirect calls are calls for which the
@@ -57,8 +55,6 @@ import com.oracle.truffle.api.frame.VirtualFrame;
  */
 public abstract class IndirectCallNode extends Node {
 
-    @Deprecated @CompilationFinal private VirtualFrame dummyFrame;
-
     /**
      * Constructor for implementation subclasses.
      *
@@ -70,30 +66,12 @@ public abstract class IndirectCallNode extends Node {
     /**
      * Performs an indirect call to the given {@link CallTarget} target with the provided arguments.
      *
-     * @param frame the caller frame
-     * @param target the {@link CallTarget} to call
-     * @param arguments the arguments to provide
-     * @return the return value of the call
-     * @since 0.8 or earlier
-     * @deprecated use call without frame instead
-     */
-    @Deprecated
-    public Object call(VirtualFrame frame, CallTarget target, Object[] arguments) {
-        return call(target, arguments);
-    }
-
-    /**
-     * Performs an indirect call to the given {@link CallTarget} target with the provided arguments.
-     *
      * @param target the {@link CallTarget} to call
      * @param arguments the arguments to provide
      * @return the return value of the call
      * @since 0.23
      */
-    public Object call(CallTarget target, Object[] arguments) {
-        // TODO change to varargs as soon as #call(VirtualFrame, Object[] will removed.
-        return call(DirectCallNode.LEGACY_FRAME, target, arguments);
-    }
+    public abstract Object call(CallTarget target, Object... arguments);
 
     /** @since 0.8 or earlier */
     public static IndirectCallNode create() {